🐛 fix: Resolve "Application error: client-side exception" in my-page
- Create modular skeleton UI components to prevent client-side rendering errors
- Remove unnecessary client state management from bookmarks layout (290→12 lines)
- Replace dynamic import wrapper with direct MyPlantsClient import
- Implement centralized LoadingRouter for stable route-based skeleton rendering
- Eliminate hydration mismatches by removing useState/useEffect in layouts
- Fix plant registration refresh using router.refresh() instead of window.reload()
- Extract BookmarksNavigation component for better separation of concerns

Technical improvements:
- Skeleton UI: LoadingRouter.tsx routes to specialized components
- Layout optimization: Remove 290-line client-side logic from bookmarks
- Component cleanup: Delete MyPlantsClientWrapper.tsx dynamic import
- Import organization: Standardize import order in MyPlantsClient
- Error prevention: Use Suspense boundaries with proper fallbacks
Ref #77
